+++ CHANGELOG.md
@@ -24,6 +24,7 @@
     * Improved support for python standard io streams
     * Exposed bitmap tracing and color quantization utilities
     * Exposed single frame rendering capabilities (raster and svg)
+    * New snippet system to run custom scripts without the need of a plugin
 * Bug Fixes:
     * Custom keyboard shortcuts are saved properly
     * Fixed values blowing up when changed from the property tree